API Gateaway/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
python-dotenv
